now we use 16bit wasm opcodes

+++ core/vm/tracer.go
@@ -2,7 +2,7 @@ package vm
 
 type OpCodeInfo interface {
 	String() string
-	Code() byte
+	Code() uint16
 	GetParams() []uint64
 	Pc() uint64
 }
